Start using cimpress-translations-cli.
