+++ package.json
@@ -4,7 +4,9 @@
   "description": "Visualizes comment(s) for a particular platform resource",
   "main": "./lib/index.js",
   "scripts": {
-    "translate": "node scripts/retrieveTranslations.js",
+    "translate": "npm run translate-i18n && npm run translate-react-timeago",
+    "translate-react-timeago": "node scripts/translate-react-timeago.js",
+    "translate-i18n": "cimpress-translations-cli -a 60eff5ed-9ddf-483e-b60a-3135c3c87435 -c ${CLIENT_ID} -s ${CLIENT_SECRET} -t ./src/locales/translations.json",
     "check": "./node_modules/.bin/eslint --ext .js,.jsx src/",
     "build": "babel src -d lib --copy-files",
     "start": "cross-env LOCAL_DEVELOPMENT=yes node_modules/.bin/start-storybook -p 3000",
@@ -44,7 +46,7 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "backstopjs": "^3.0.39",
-    "cimpress-translations": "^2.0.2",
+    "cimpress-translations-cli": "^0.1.4",
     "core-js": "^2.5.7",
     "cross-env": "^5.2.0",
     "eslint": "^5.3.0",
