--- src/locales/translations.eng.json
@@ -1,21 +1,0 @@
-{
-    "translations": {
-        "show_my_comment_to": "Show my comment to",
-        "use_at_char_for_mentions": "Use <0>@</0> character to mention people. We will notify them via email about this comment thread.",
-        "visibility_option_internal_label": "Only authorized people",
-        "btn_post": "Post",
-        "Forbidden": "Forbidden",
-        "unable_to_retrieve_comments": "Unable to retrieve comments.",
-        "unable_to_read_comment": "Unable to retrieve comment.",
-        "retrieving_comments": "Retrieving comments...",
-        "no_comments_exist": "No comments here yet.",
-        "visibility_option_internal_description": "Those who have access to the resource I'm commenting on",
-        "visibility_option_public_description": "Every person in Cimpress",
-        "by": "by",
-        "modified": "modified",
-        "unable_to_post_comment": "Unable to post comment",
-        "visibility_option_public_label": "Everyone",
-        "btn_close": "Close",
-        "header_comments": "Comments"
-    }
-}
