--- src/locales/translations.fra.json
@@ -1,21 +1,0 @@
-{
-    "translations": {
-        "show_my_comment_to": "Montrer mon commentaire à",
-        "use_at_char_for_mentions": "Utilisez le caractère <0>@</0> pour mentionner les personnes. Nous les informerons par e-mail de ce sujet.",
-        "visibility_option_internal_label": "Seules les personnes autorisées",
-        "btn_post": "Poster",
-        "Forbidden": "Interdit",
-        "unable_to_retrieve_comments": "Impossible de récupérer des commentaires.",
-        "unable_to_read_comment": "Impossible de récupérer le commentaire.",
-        "retrieving_comments": "Récupération des commentaires ...",
-        "no_comments_exist": "Pas encore de commentaires ici.",
-        "visibility_option_internal_description": "Ceux qui ont accès à la ressource que je commente",
-        "visibility_option_public_description": "Chaque personne dans Cimpress",
-        "by": "par",
-        "modified": "modifié",
-        "unable_to_post_comment": "Impossible de poster un commentaire",
-        "visibility_option_public_label": "Toutes les personnes",
-        "btn_close": "Fermer",
-        "header_comments": "Commentaires"
-    }
-}
