--- src/locales/translations.pol.json
@@ -1,21 +1,0 @@
-{
-    "translations": {
-        "show_my_comment_to": "Kto może zobaczyć mój komentarz",
-        "use_at_char_for_mentions": "Użyj znaku <strong>@</ strong>, by wspomnieć konkretne osoby w swoim komentarzu, a powiadomimy je o nim przy pomocy e-mail'a.",
-        "visibility_option_internal_label": "Tylko użytkownicy z uprawnieniami do zasobu",
-        "btn_post": "Wyślij",
-        "Forbidden": "Dostęp zabroniony",
-        "unable_to_retrieve_comments": "Nie udało się pobrać komentarzy.",
-        "unable_to_read_comment": "Nie można pobrać komentarzy.",
-        "retrieving_comments": "Trwa pobieranie komentarzy ...",
-        "no_comments_exist": "Nie ma jeszcze żadnych komentarzy.",
-        "visibility_option_internal_description": "Użytkownicy mający dostęp do odczytywania zasobu który jest przedmiotem komentowania",
-        "visibility_option_public_description": "Każda osoba posiadająca konto na platformie",
-        "by": "przez",
-        "modified": "zmodyfikowany",
-        "unable_to_post_comment": "Nie można opublikować komentarza",
-        "visibility_option_public_label": "Wszyscy",
-        "btn_close": "Zamknij",
-        "header_comments": "Komentarze"
-    }
-}
