--- src/locales/translations.spa.json
@@ -1,21 +1,0 @@
-{
-    "translations": {
-        "show_my_comment_to": "Mostrar mi comentario a",
-        "use_at_char_for_mentions": "Usa el personaje <strong> @ </ strong> para mencionar a las personas. Los notificaremos por correo electrónico sobre este hilo de comentarios.",
-        "visibility_option_internal_label": "Solo personas autorizadas",
-        "btn_post": "Enviar",
-        "Forbidden": "Prohibido",
-        "unable_to_retrieve_comments": "No se pueden recuperar los comentarios.",
-        "unable_to_read_comment": "No se puede recuperar el comentario.",
-        "retrieving_comments": "Recuperando comentarios ...",
-        "no_comments_exist": "No hay comentarios aquí todavía.",
-        "visibility_option_internal_description": "Aquellos que tienen acceso al recurso que estoy comentando",
-        "visibility_option_public_description": "Cada persona en Cimpress",
-        "by": "por",
-        "modified": "modificado",
-        "unable_to_post_comment": "No disponible para dejar comentario",
-        "visibility_option_public_label": "Todo el mundo",
-        "btn_close": "Cerca",
-        "header_comments": "Comentarios"
-    }
-}
